Almost works

+++ webservice/src/main/kotlin/com/noticemedan/finch/dto/ttn/LoraInfo.kt
@@ -4,5 +4,5 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 data class LoraInfo (
 		@JsonProperty("bandwidth") val bandwidth: Long,
-		@JsonProperty("spreading_factor") val spreadingFactor: Int
+		@JsonProperty("spreading_factor") val spreadingFactor: Long
 )
