+++ webservice/src/main/kotlin/com/noticemedan/finch/dto/ttn/UplinkMessageInfo.kt
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import java.time.LocalDateTime
 
 data class UplinkMessageInfo (
-		@JsonProperty("f_port") val fPort: Int,
-		@JsonProperty("f_cnt") val frameCount: Int,
+		@JsonProperty("f_port") val fPort: Long,
+		@JsonProperty("f_cnt") val frameCount: Long,
 		@JsonProperty("frm_payload") val framePayload: String,
 		@JsonProperty("rx_metadata") val rxMetadata: List<RxMetadataInfo>,
 		@JsonProperty("settings") val settings: SettingsInfo,
